Refactor scripts directory structure
