+++ knip.config.ts
@@ -1,5 +1,5 @@
 import type { KnipConfig } from 'knip'
-import { getIgnores } from './scripts/ignores.js'
+import { getIgnores } from './scripts/utils/ignores.js'
 
 const ADDITIONALLY_ENTRY_POINTS = [
 	'src/routes/sayno/SelfieUX.svelte', // dynamically imported
