+++ scripts/l10n/utils/additions.ts
