+++ scripts/l10n/utils/branch-safety.test.ts
