+++ scripts/l10n/utils/branch-safety.ts
