+++ scripts/l10n/utils/dry-run.ts
