+++ scripts/l10n/utils/force.test.ts
