+++ scripts/l10n/utils/force.ts
