+++ scripts/l10n/utils/heart.ts
