+++ scripts/l10n/utils/llm-utils.ts
