+++ scripts/l10n/utils/mode.test.ts
