+++ scripts/l10n/utils/prompts.ts
