+++ scripts/l10n/utils/utils.ts
