+++ scripts/local/open-letter/fetch-scottish-portraits.js
