+++ scripts/local/open-letter/process-portraits.js
