+++ scripts/utils/ignores.js
